Synthesizing primitive inspections
