+++ src/Language/Mulang/Analyzer/ExpectationsCompiler.hs
@@ -1,3 +1,5 @@
+{-# LANGUAGE ViewPatterns #-}
+
 module Language.Mulang.Analyzer.ExpectationsCompiler(
   compileExpectation) where
 
@@ -99,10 +101,10 @@ compileInspectionPrimitive = f
   f ["Instantiates"]                   = bound instantiates
   f ["Raises"]                         = bound raises
   f ["Rescues"]                        = bound rescues
+  f ("Returns":args)                   = plain (returnsMatching (compileMatcher args))
   f ["TypesAs"]                        = bound typesAs
   f ["TypesParameterAs"]               = bound typesParameterAs
   f ["TypesReturnAs"]                  = bound typesReturnAs
-  f ("Returns":args)                   = plain (returnsMatching (compileMatcher args))
   f ["Uses"]                           = bound uses
   f ["UsesAnonymousVariable"]          = plain usesAnonymousVariable
   f ["UsesBooleanLogic"]               = plain usesBooleanLogic
@@ -128,6 +130,7 @@ compileInspectionPrimitive = f
   f ["UsesNot"]                        = plain usesNot
   f ["UsesObjectComposition"]          = plain usesObjectComposition
   f ["UsesPatternMatching"]            = plain usesPatternMatching
+  f ["UsesPrint"]                      = plain usesPrint
   f ["UsesRepeat"]                     = plain usesRepeat
   f ["UsesStaticMethodOverload"]       = plain usesStaticMethodOverload
   f ["UsesStaticPolymorphism"]         = contextualized usesStaticPolymorphism'
@@ -136,7 +139,12 @@ compileInspectionPrimitive = f
   f ["UsesType"]                       = bound usesType
   f ["UsesWhile"]                      = plain usesWhile
   f ["UsesYield"]                      = plain usesYield
+  f [primitiveUsage -> Just p]         = plain (usesPrimitive p)
+  f [primitiveDeclaration -> Just p]   = plain (declaresPrimitive p)
   f _                                  = Nothing
+
+  primitiveUsage = decodeUsageInspection
+  primitiveDeclaration = decodeDeclarationInspection
 
   contextualized :: ContextualizedInspection -> Maybe ContextualizedBoundInspection
   contextualized = Just . contextualizedBind
